pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "street-view-mcp"
version = "0.1.0"
description = "MCP for interacting with Street View imagery"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "requests",
    "pillow",
    "python-dotenv",
    "fastmcp",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/street-view-mcp"
"Bug Tracker" = "https://github.com/yourusername/street-view-mcp/issues"

[project.scripts]
street-view-mcp = "street_view_mcp.street_view:main"
street-view-mcp-server = "street_view_mcp.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/street_view_mcp"]

[tool.hatch.envs.default]
dependencies = [
    "pytest",
    "pytest-cov",
]
